add crew list to movies
This commit adds a crew list to each movie. It
allows filtering of the list by crew name and job.
Any crew members with multiple jobs are tweaked
to store as a single crew member with a list of
jobs.

This commit also removes unnecessary code from
the previous commits.

Signed-off-by: Rebecca Troy <20088744@mail.wit.ie>